requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.7
django-cors-headers==3.1.1
django-filter==2.2.0
djangorestframework==3.10.3
Pillow==6.2.1
python-decouple==3.1
pytz==2019.3
sqlparse==0.3.0
static3==0.7.0
gunicorn
psycopg2
